requirements.txt
discord.py
python-dotenv~=0.19.2